requirements.txt
aiohttp==3.7.4.post0
akinator.py==5.0.0
astunparse==1.6.3
async-timeout==3.0.1
asyncdagpi==3.3.4
asyncio==3.4.3
attrs==21.2.0
braceexpand==0.1.7
brotlipy==0.7.0
buttons==0.1.9
certifi==2021.5.30
chardet==3.0.4
colorama==0.4.4
discord==1.7.3
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@fbb8803779373357e274e1540b368365fd9d8074
discord.py @ git+https://github.com/Rapptz/discord.py@3382d2e9e86f4daacd9097f284da51b49a54f7b9
dnspython==2.1.0
fuzzywuzzy==0.18.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httpx==0.13.3
humanize==3.10.0
hyperframe==5.2.0
idna==2.10
import-expression==1.1.4
jishaku==2.1.0
mazelib==0.9.12
motor==2.4.0
multidict==5.1.0
numpy==1.20.1
Pillow==8.3.1
pymongo==3.11.4
randomstuff.py==2.0.0
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
thefuzz==0.19.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3